refactor: refactor sync providers (#291)
## This PR

Revisits sync providers and refactor them.

- unify event handling
- packaging improvements
- removal of unused codes (cleanup)
- clean contracts between Runtime and Sync implementations 

With this change, we get a clear isolation between runtime and sync
providers (ex:- file, k8s, ....). For example, the runtime is not aware
of any sync implementation details such as events. It simply coordinates
sync impls, store (evaluator) and notifications. This should bring more
extensibility when adding future extensions.

Below is the overview of the internals and interactions,


![image](https://user-images.githubusercontent.com/8186721/213037032-316adb7e-e9ab-42e3-82f6-c3eaa2612ba3.png)

Signed-off-by: Kavindu Dodanduwa <kavindudodanduwa@gmail.com>