fix: fix race in http sync test (#401)
In the `TestSimpleSync` test we are asserting that `hs.Cron.AddFunc` and
`hs.Cron.Start` are called by the time the `dataSync` is sent. This is a
race, because these methods are actually called after the first send. I
get failures here ~25% of the time (seems like @skyerus does too).

This changes things to add the handler and start the cron _before_ the
first send, which makes things deterministic.

Signed-off-by: Todd Baert <toddbaert@gmail.com>