docs: add json-schema (#1115)
Adds JSON schema at
`https://flagd.dev/schema/v0/flagd-definitions.json`, which can be used
to add validation to flagd config files.

~Note, to accomplish this cleanly, I have cloned the schemas module
twice (at different versions) one is used for json releases (and tagged
to something like `json/json-schema-v0.1.2`) while the other is for the
protobuf generation and tagged to a release of that. I don't love this,
but I think it's needed.~

I've decided against this. I don't think we'll really have conflicts
here.

See schema here:
https://deploy-preview-1115--polite-licorice-3db33c.netlify.app/schema/v0/flagd-definitions.json

Once this is available, I intend to update a bunch of docs and examples
to leverage this schema, and and an blurb in the docs about it.

~Blocked by: https://github.com/open-feature/flagd-schemas/pull/122~

~I think I'll wait for
[this](https://github.com/open-feature/flagd/pull/1146) to be done
before adding this, to avoid conflicts.~

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>
Co-authored-by: Michael Beemer <beeme1mr@users.noreply.github.com>