fix: incorrect metadata returned per source (#1599)
This PR fixes an issue where the wrong metadata is returned when a
selector is used.

This bug is easy to fix, but was also easy to write. We eventually need
to refactor the flag memory store (`flags.go`) to store all flags per
source. This was retrofitted in to support selectors, from a
single-flagSet model.

To test, do the manual test specified in
https://github.com/open-feature/flagd/issues/1597.

Fixes: https://github.com/open-feature/flagd/issues/1597

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>