refactor: removed unused Selector from Flag and Store. (#1747)
## This PR

This pull request removes the unused `Selector` field from the `Flag`
struct and deletes the `SourceDetails` struct, which is no longer in
use.

These changes are part of a cleanup effort following the major
refactoring in PR #1702, which made these components obsolete.

### Related Issues

Follow-up to #1702

### How to test

All existing tests should continue to pass.

Signed-off-by: Andrey <andreyturkov@google.com>
Co-authored-by: Todd Baert <todd.baert@dynatrace.com>