test: add `buildHeaders` function for http sync tests (#1804)
## This PR

The previous way of building HTTP headers with raw map is not reliable
in some edge cases.

An example is case-sensitive issue for `ETag` header which requires
`Etag` (lower case `t`) to be the map key when building the response
with a raw map, in order to receive it via `resp.Header.Get("ETag")` on
the other end.

Having a `buildHeaders` function to utilize `Set` built-in function of
`http.Header` would solve this concern.

### How to test
```bash
go test ./core/pkg/sync/http/...
```

---------

Signed-off-by: Zhiwei Liang <zhiwei.liang27@pm.me>